requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
chardet==4.0.0
click==8.1.7
distro==1.9.0
dnspython==1.16.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
Flask==2.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.0
openai==1.35.10
openpyxl==3.1.5
pandas==2.2.2
pandas-stubs==2.2.2.240603
pydantic==2.8.2
pydantic_core==2.20.1
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==0.17.1
pytz==2024.1
requests==2.25.1
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
types-pytz==2024.1.0.20240417
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
Werkzeug==2.3.7
